+++ src/main/java/com/moilioncircle/redis/replicator/rdb/dump/DumpRdbVisitor.java
@@ -333,7 +333,7 @@ public class DumpRdbVisitor extends DefaultRdbVisitor {
         replicator.addRawByteListener(listener);
         SkipRdbParser skipParser = new SkipRdbParser(in);
         long len = skipParser.rdbLoadLen().len;
-        for (int i = 0; i < len; i++) {
+        for (long i = 0; i < len; i++) {
             skipParser.rdbGenericLoadStringObject();
         }
         replicator.removeRawByteListener(listener);
